Trocando long e int por Integer

+++ src/main/java/com/sistema/apicr7imports/controller/BrandController.java
@@ -36,7 +36,7 @@ public class BrandController {
 
 	@ApiOperation(value = "Trazer tipo de marca cadastrada por id")
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Brand> findById(@ApiParam(value = "ID de Cadastro no Banco", required = true, example = "1") @PathVariable Long id) {
+	public ResponseEntity<Brand> findById(@ApiParam(value = "ID de Cadastro no Banco", required = true, example = "1") @PathVariable Integer id) {
 		return ResponseEntity.ok().body(service.findbyId(id));
 	}
 
@@ -48,7 +48,7 @@ public class BrandController {
 
 	@ApiOperation(value = "Deleta uma marca")
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> delete(@ApiParam(value = "ID de Cadastro no Banco", required = true, example = "1") @PathVariable Long id) {
+	public ResponseEntity<Void> delete(@ApiParam(value = "ID de Cadastro no Banco", required = true, example = "1") @PathVariable Integer id) {
 		service.delete(id);
 		return ResponseEntity.noContent().build();
 	}
